Fixed compiler warning on line 451 of plfs_fuse.cpp
git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@821 979fb63a-54b4-4514-8d67-9ab8ea1b59df

+++ fuse/plfs_fuse.cpp
@@ -448,7 +448,12 @@ int Plfs::f_access(const char *path, int mask) {
 int Plfs::f_mknod(const char *path, mode_t mode, dev_t rdev) {
     PLFS_ENTER;
 
-    mlog(FUSE_DAPI, "%s on %s mode %d rdev %d",__FUNCTION__,path,mode,rdev);
+    mlog(FUSE_DAPI,
+            "%s on %s mode %d rdev %d",
+            __FUNCTION__,
+            path,
+            mode,
+            (int)rdev);
 
     ret = makePlfsFile( strPath.c_str(), mode, 0 );
     if ( ret == 0 ) {
